pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "fkcrypto"
version = "0.1.0"
description = "Crypto Trading Agent System — Hybrid multi-agent architecture with Freqtrade integration"
readme = "README.md"
requires-python = ">=3.11"
license = {text = "MIT"}
authors = [
    {name = "FKCrypto Team"}
]
dependencies = [
    # Core
    "langgraph>=0.2.0",
    "langchain>=0.3.0",
    "langchain-openai>=0.2.0",
    "langchain-anthropic>=0.2.0",
    "litellm>=1.50.0",
    # Data
    "ccxt>=4.0.0",
    "pycoingecko>=3.1.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "ta>=0.11.0",
    # Database
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    # Message queue
    "redis>=5.0.0",
    # Scheduling
    "apscheduler>=3.10.0",
    # API
    "fastapi>=0.110.0",
    "uvicorn[standard]>=0.29.0",
    "httpx>=0.27.0",
    # Config
    "pyyaml>=6.0",
    "python-dotenv>=1.0.0",
    "pydantic>=2.0.0",
    "pydantic-settings>=2.0.0",
    # Notifications
    "python-telegram-bot>=21.0",
    "aiohttp>=3.9.0",
    # Observability
    "structlog>=24.0.0",
    "prometheus-client>=0.20.0",
    # Utils
    "python-dateutil>=2.9.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=5.0",
    "ruff>=0.4.0",
    "mypy>=1.10",
    "pre-commit>=3.0",
]
dashboard = [
    "streamlit>=1.35.0",
    "plotly>=5.22.0",
]

[project.scripts]
fkcrypto = "src.gateway:main"

[tool.setuptools.packages.find]
include = ["src*"]

[tool.ruff]
target-version = "py311"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.mypy]
python_version = "3.11"
strict = true
warn_return_any = true
warn_unused_configs = true
